Refactor: python, cts: Move RemoteStonithd into the _cts module.
No code changes.
